[AAE-5909] added story for json cell component, separate and improve datatable structure for json cell usage
