+++ src/app/stories/datatable/datatable-custom-data/datatable-custom-data.component.html
@@ -1,1 +1,1 @@
-<adf-datatable [columns]="columns" [rows]="rows" [sorting]="sorting" [data]="data" [resolverFn]="resolverFn"> </adf-datatable>
+<adf-datatable [columns]="columns" [rows]="rows" [sorting]="sorting" [data]="data" [resolverFn]="resolverFn"></adf-datatable>
