+++ src/app/stories/datatable/datatable.component.html
@@ -37,5 +37,5 @@
       </ng-template>
     </adf-loading-content-template>
   </adf-datatable>
-  <h1 *ngIf="noPermission" style="text-align: center;">You don't have permissions</h1>
+  <h1 *ngIf="noPermission" style="text-align: center">You don't have permissions</h1>
 </div>
