+++ src/app/stories/json-cell/json-cell.component.html
@@ -1,10 +1,9 @@
 <adf-json-cell
-  [column]="column"
-  [copyContent]="copyContent"
   [data]="data"
   [editable]="editable"
-  [resolverFn]="resolverFn"
+  [copyContent]="copyContent"
   [row]="row"
   [tooltip]="tooltip"
->
-</adf-json-cell>
+  [column]="column"
+  [resolverFn]="resolverFn"
+></adf-json-cell>
