+++ src/app/stories/json-cell/json-cell.stories.ts
@@ -5,6 +5,7 @@ import { RouterModule } from '@angular/router';
 import { Meta, moduleMetadata, Story } from '@storybook/angular';
 import { APP_ROUTES } from '../../app.routes';
 import { AppCommonModule } from '../../components/common/common.module';
+import { FilteredDataAdapter } from '../datatable/datatable-custom-data/filtered-data-adapter';
 import { JsonCellComponent } from './json-cell.component';
 
 export default {
@@ -25,7 +26,14 @@ export default {
         BrowserAnimationsModule
       ]
     })
-  ]
+  ],
+  argTypes: {
+    column: { table: { disable: true } },
+    data: { table: { disable: true } },
+    copyContent: { table: { disable: true } },
+    tooltip: { table: { disable: true } },
+    resolverFn: { table: { disable: true } }
+  }
 } as Meta;
 
 const Template: Story<JsonCellComponent> = (args) => ({
@@ -34,13 +42,24 @@ const Template: Story<JsonCellComponent> = (args) => ({
   }
 });
 
-export const DefaultStory = Template.bind({});
-DefaultStory.args = {
-  column: null,
+const resolverData = new FilteredDataAdapter();
+
+const rows = [
+  { id: 1, name: '1 Name 1' },
+  { id: 2, name: '2 Name 2' },
+  { id: 3, name: '3 Name 3' },
+  { id: 4, name: '4 Name 4' }
+];
+
+const columnsSchema = [];
+
+export const EditableJSON = Template.bind({});
+EditableJSON.args = {
+  column: columnsSchema,
   copyContent: true,
-  data: null,
-  editable: false,
+  data: resolverData,
+  editable: true,
+  row: rows,
   resolverFn: null,
-  row: null,
   tooltip: 'JSON cell'
 };
